Объеденит доч компоненты HeaderNav в один компонент
