+++ project/src/index.tsx
@@ -20,12 +20,10 @@ ReactDOM.render(
   document.getElementById('root'),
 );
 
-//Упростить обработчик сортировки в onOffersSortTypeChange
-
-//Вынести Header их отрисовки
+//Доработать Header
 //Разбить редьюсер
+//Упростить обработчик сортировки в onOffersSortTypeChange
 //Настроить оптимизацию
-//Пройтись по критериям, отрефакторить код, вынести повторяющияйся код в функции
 
 //Оптимизация /
 //Оптимизировать перерисовку Header Navigation
@@ -36,9 +34,7 @@ ReactDOM.render(
 //Оптимизировать рендеринг текста при заполнении рейтинга
 
 //Оптимизация /favorites
-//Оптимизировать рендеринг Rating при заполнениии text
-//Оптимизировать рендеринг текста при заполнении рейтинга
 
 //Оптимизация /login
-//Оптимизировать рендеринг Rating при заполнениии text
-//Оптимизировать рендеринг текста при заполнении рейтинга
+
+//Пройтись по критериям, отрефакторить код, вынести повторяющияйся код в функции
