--- project/src/pages/header-login/header-login.tsx
@@ -1,15 +1,0 @@
-import { Link } from 'react-router-dom';
-import { AppRoute } from '../../const';
-
-function HeaderLogin(): JSX.Element {
-  return (
-    <li className="header__nav-item user">
-      <Link to={AppRoute.Login} className="header__nav-link header__nav-link--profile">
-        <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-        <span className="header__login">Sign in</span>
-      </Link>
-    </li>
-  );
-}
-
-export default HeaderLogin;
